Add samples from  Direct3D9 tutorial

+++ source/Graphics/DirectX9/Direct3D/TextureImpl.cpp
@@ -24,7 +24,7 @@ TextureImpl::TextureImpl(RenderContextImpl* renderContextImpl, const Point2u& si
 	else
 		format = D3DFMT_R8G8B8;
 	
-	result = D3DXCreateTexture(renderContextImpl->Context.Device, size.PosX(), size.PosY(), D3DX_DEFAULT, 0, format, D3DPOOL_MANAGED, &_Texture);
+	result = D3DXCreateTexture(renderContextImpl->Context.Device, (UINT)size.PosX(), (UINT)size.PosY(), D3DX_DEFAULT, 0, format, D3DPOOL_MANAGED, &_Texture);
 	if (FAILED(result))
 		throw LDL::Core::RuntimeError("D3DXCreateTexture failed");
